Got my parser working!!! woot woot!! It will accept a variety of real life ingredient strings and parse them into appropriate fields. I'm still dependent on specific formatting, but, mostly same as normal recipes.
Examples:
1 28-oz can tomatoes, drained, juice reserved
5 small tart apples
salt, to taste
pepper
1 large red onion
1 c low-sodium chicken broth
4 whole-wheat pitas
12 boneless skinless chicken thighs

Note: if you use 'whole' you'll have to hyphenate it to the following word, because 'whole' is a
unit amount. for now, anyways. I may later simply assume that if there is no unit, it's the whole thing.
If you use a package size & unit, such as 28-oz, it needs to be hyphenated.
Everything after the first comma is considered to be an additional note/pre-prep, whatever.